修复数组/元组/结构体比较 (#33)
* 修复索引赋值

* 修复比较

* 修复test

+++ test/return_test.go
@@ -6,7 +6,7 @@ import (
 
 func TestReturn(t *testing.T) {
 	assertRetEqZero(t, `
-func main()isize{
+func main()u8{
     return 0
 }
 `)
